requirements.txt
easyocr==1.7.1
numpy==1.23.2
opencv_contrib_python==4.10.0.84
Pillow==10.4.0
streamlit==1.38.0
ultralytics==8.2.82
